output of quadrature encoder data over serial
This is an arduino program written in wiring that was tested to monitor a 600 pulse-per-revolution output of a quadrature type of rotary encoder. It will work with any type of quadrature rotary encoder that is of the Taiss E38S6 type, 4 wire rotary encoder. It was written for Green-OUTA, White-OUTB, Red-VCC, Black-0V and assembled into an Arduino nano.